feature: add merge tus file (#237)
* fix: :bug: correct CORS exposed headers formatting in SecurityConfig

* fix: :bug: update Tus extension header value in TusController

* fix: :bug: update Tus file location handling to use configurable path prefix

* fix: :bug: change uploadLength type to Integer and improve metadata parsing error handling

* fix: :bug: update mergePath configuration and adjust file merging logic in TusService

+++ src/main/java/com/springboot/api/tus/entity/TusFileInfo.java
@@ -69,7 +69,7 @@ public class TusFileInfo extends BaseEntity {
         super.onCreate();
     }
 
-    public void updateOffset(Long uploadLength) {
+    public void updateOffset(Integer uploadLength) {
         contentOffset += uploadLength;
     }
